Updated from comments on PR

+++ server/server.py
@@ -5,17 +5,45 @@ import os
 from RAGHelper_cloud import RAGHelperCloud
 from RAGHelper_local import RAGHelperLocal
 from pymilvus import Collection, connections
+from werkzeug.utils import secure_filename
 
-load_dotenv()
 
+def load_bashrc():
+    """
+    Load environment variables from the user's .bashrc file.
+
+    This function looks for the .bashrc file in the user's home directory
+    and loads any environment variables defined with the 'export' command
+    into the current environment.
+    """
+    bashrc_path = os.path.expanduser("~/.bashrc")
+    if os.path.exists(bashrc_path):
+        with open(bashrc_path) as f:
+            for line in f:
+                if line.startswith("export "):
+                    key, value = line.strip().replace("export ", "").split("=", 1)
+                    value = value.strip(' "\'')
+                    os.environ[key] = value
+
+
+# Initialize Flask application
 app = Flask(__name__)
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-# Instantiate the RAG Helper class
-if os.getenv("use_openai") == "True" or os.getenv("use_gemini") == "True" or os.getenv("use_azure") == "True":
+# Disable parallelism in tokenizers to avoid warnings
+os.environ["TOKENIZERS_PARALLELISM"] = "false"
+
+# Load environment variables from .env file
+load_bashrc()
+load_dotenv()
+
+# Instantiate the RAG Helper class based on the environment configuration
+if any(os.getenv(key) == "True" for key in ["use_openai", "use_gemini", "use_azure", "use_ollama"]):
+    logger.info("Instantiating the cloud RAG helper.")
     raghelper = RAGHelperCloud(logger)
 else:
+    logger.info("Instantiating the local RAG helper.")
     raghelper = RAGHelperLocal(logger)
 
 
@@ -30,13 +58,27 @@ def add_document():
     Returns:
         JSON response with the filename and HTTP status code 200.
     """
-    json = request.get_json()
-    filename = json['filename']
+    if 'file' not in request.files:
+        return jsonify({"error": "No file part in the request"}), 400
 
-    raghelper.addDocument(filename)
+    file = request.files['file']
+    if file.filename == '':
+        return jsonify({"error": "No file selected"}), 400
+
+    if file:
+        filename = secure_filename(file.filename)
+
+        logger.info(f"Adding document {filename}")
+        data_dir = os.getenv("data_directory")
+        location = f"{data_dir}/{file.filename}"
+
+        # Copy over file
+        file.save(location)
+        raghelper.add_document(location)
+    else:
+        return jsonify({"error": "File is required"}), 400
 
     return jsonify({"filename": filename}), 200
-
 
 @app.route("/query", methods=['POST'])
 def query():
@@ -62,7 +104,6 @@ def query():
         logger.error(f"Error processing query: {str(e)}")
         return jsonify({'error': str(e)}), 500
 
-
 @app.route("/chat", methods=['POST'])
 def chat():
     """
@@ -74,14 +115,16 @@ def chat():
     Returns:
         JSON response containing the assistant's reply, history, documents, and other metadata.
     """
-    json = request.get_json()
-    prompt = json['prompt']
-    history = json.get('history', [])
-    original_docs = json.get('docs', [])
+    json_data = request.get_json()
+    prompt = json_data.get('prompt')
+    history = json_data.get('history', [])
+    original_docs = json_data.get('docs', [])
     docs = original_docs
 
     # Get the LLM response
     (new_history, response) = raghelper.handle_user_interaction(prompt, history)
+    if not docs or 'docs' in response:
+        docs = response['docs']
 
     # Handle citation verification errors if any
     if response.get('error') == 'citation_verification_failed':
@@ -92,44 +135,56 @@ def chat():
 
     if len(docs) == 0 or 'docs' in response:
         docs = response['docs']
-    # Break up the response for OS LLMs
+
+    # Break up the response for local LLMs
     if isinstance(raghelper, RAGHelperLocal):
         end_string = os.getenv("llm_assistant_token")
         reply = response['text'][response['text'].rindex(end_string) + len(end_string):]
 
-        # Get history
+        # Get updated history
         new_history = [{"role": msg["role"], "content": msg["content"].format_map(response)} for msg in new_history]
         new_history.append({"role": "assistant", "content": reply})
     else:
-        # Populate history properly, also turning it into dict instead of tuple, so we can serialize
+        # Populate history for other LLMs
         new_history = [{"role": msg[0], "content": msg[1].format_map(response)} for msg in new_history]
         new_history.append({"role": "assistant", "content": response['answer']})
         reply = response['answer']
-    # Make sure we format the docs properly
 
-    if len(original_docs) == 0 or 'docs' in response:
+    # Format documents
+    fetched_new_documents = False
+    if not original_docs or 'docs' in response:
+        fetched_new_documents = True
         new_docs = [{
             's': doc.metadata['source'],
             'c': doc.page_content,
             **({'pk': doc.metadata['pk']} if 'pk' in doc.metadata else {}),
-            **({'provenance': float(doc.metadata['provenance'])} if 'provenance' in doc.metadata else {}),
+            **({'provenance': float(doc.metadata['provenance'])} if 'provenance' in doc.metadata and doc.metadata[
+                'provenance'] is not None else {}),
             **({'citation_verification': response.get('citation_verification',
                                                       {})} if 'citation_verification' in response else {})
         } for doc in docs if 'source' in doc.metadata]
     else:
         new_docs = docs
 
-    result = {
+    # Build the response dictionary
+    response_dict = {
         "reply": reply,
         "history": new_history,
-        "documents": new_docs
+        "documents": new_docs,
+        "rewritten": False,
+        "question": prompt,
+        "fetched_new_documents": fetched_new_documents
     }
 
+    # Check for rewritten question
+    if os.getenv("use_rewrite_loop") == "True" and prompt != response['question']:
+        response_dict["rewritten"] = True
+        response_dict["question"] = response['question']
     # Add citation verification results if present
     if 'citation_verification' in response:
         result['citation_verification'] = response['citation_verification']
 
-    return jsonify(result), 200
+    return jsonify(response_dict), 200
 
 
 @app.route("/get_documents", methods=['GET'])
@@ -144,9 +199,12 @@ def get_documents():
         JSON response containing the list of files.
     """
     data_dir = os.getenv('data_directory')
-    file_types = os.getenv("file_types").split(",")
-    files = [f for f in os.listdir(data_dir) if
-             os.path.isfile(os.path.join(data_dir, f)) and os.path.splitext(f)[1][1:] in file_types]
+    file_types = os.getenv("file_types", "").split(",")
+
+    # Filter files based on specified types
+    files = [f for f in os.listdir(data_dir)
+             if os.path.isfile(os.path.join(data_dir, f)) and os.path.splitext(f)[1][1:] in file_types]
+
     return jsonify(files)
 
 
@@ -162,8 +220,8 @@ def get_document():
         JSON response with the error message and HTTP status code 404 if not found,
         otherwise sends the file as an attachment.
     """
-    json = request.get_json()
-    filename = json['filename']
+    json_data = request.get_json()
+    filename = json_data.get('filename')
     data_dir = os.getenv('data_directory')
     file_path = os.path.join(data_dir, filename)
 
@@ -187,13 +245,16 @@ def delete_document():
     Returns:
         JSON response with the count of deleted documents.
     """
-    json = request.get_json()
-    filename = json['filename']
+    json_data = request.get_json()
+    filename = json_data.get('filename')
     data_dir = os.getenv('data_directory')
     file_path = os.path.join(data_dir, filename)
 
+    if not os.path.exists(file_path):
+        return jsonify({"error": "File not found"}), 404
+
     # Remove from Milvus
-    connections.connect(uri=os.getenv('vector_store_path'))
+    connections.connect(uri=os.getenv('vector_store_uri'))
     collection = Collection("LangChainCollection")
     collection.load()
     result = collection.delete(f'source == "{file_path}"')
@@ -202,7 +263,7 @@ def delete_document():
     # Remove from disk too
     os.remove(file_path)
 
-    # BM25 needs to be re-initialized
+    # Reinitialize BM25
     raghelper.loadData()
 
     return jsonify({"count": result.delete_count})
